final stable commit, added docs, some fixes, added Author line, refactoring
