+++ src/main/java/webapp/marginality2/model/AbstractEntity.java
@@ -1,17 +1,22 @@
 package webapp.marginality2.model;
 
+import lombok.Data;
 import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import java.time.LocalDate;
-import java.util.Objects;
 
 /**
- * AbstractEntity is a base class for entity definitions in the application.
- * It provides common fields and functionality to its subclasses.
+ * AbstractEntity — это базовый класс для всех сущностей в приложении.
+ * Он предоставляет общие поля, такие как id, cost, profit, count, name, status и date,
+ * которые могут использоваться в наследуемых классах для минимизации дублирования кода.
+ *
+ * @author Ivan Pshychenko
+ * © 2024 Ivan Pshychenko ☁️
  */
-@lombok.Data
+@Data
 @NoArgsConstructor
 public abstract class AbstractEntity {
+
     @Id
     private int id;
     private int cost;
@@ -21,6 +26,17 @@ public abstract class AbstractEntity {
     private Status status;
     private LocalDate date;
 
+    /**
+     * Конструктор с полным набором параметров для инициализации экземпляра сущности.
+     *
+     * @param id     Уникальный идентификатор для сущности
+     * @param cost   Стоимость блюда или товара
+     * @param profit Прибыль от продажи блюда или товара
+     * @param count  Количество единиц товара или блюда
+     * @param name   Название блюда или товара
+     * @param status Статус товара или блюда (например, ПРОДАНО, НА СКЛАДЕ)
+     * @param date   Дата, связанная с блюдом (например, дата продажи или производства)
+     */
     protected AbstractEntity(int id, int cost, int profit, int count, String name, Status status, LocalDate date) {
         this.id = id;
         this.cost = cost;
@@ -31,6 +47,12 @@ public abstract class AbstractEntity {
         this.date = date;
     }
 
+    /**
+     * Переопределение метода equals для сравнения сущностей по полю id.
+     *
+     * @param o Объект для сравнения
+     * @return true, если оба объекта одного типа и имеют одинаковый id, иначе false
+     */
     @Override
     public boolean equals(Object o) {
         if (this == o) return true;
@@ -39,8 +61,16 @@ public abstract class AbstractEntity {
         return id == that.id;
     }
 
+    /**
+     * Улучшенная версия метода hashCode для генерации уникального хеш-кода на основе всех полей объекта.
+     *
+     * @return Хеш-код, сгенерированный на основе всех ключевых полей
+     */
     @Override
     public int hashCode() {
-        return Objects.hashCode(id);
+        int prime = 31;
+        int result = 1;
+        result = prime * result + id;
+        return result;
     }
 }
