+++ src/main/java/webapp/marginality2/model/Status.java
@@ -1,5 +1,12 @@
 package webapp.marginality2.model;
 
+/**
+ * Перечисление статусов для сущности Meal.
+ * Статусы отражают текущее состояние продукта, такого как "продано", "истек срок годности", "на продаже", "на хранении", и "повреждено".
+ *
+ * @author Ivan Pshychenko
+ * © 2024 Ivan Pshychenko ☁️
+ */
 public enum Status {
     SOLD,
     EXPIRED,
