--- src/main/resources/META-INF/resources/frontend/themes/custom-lumo-theme/custom-lumo-theme.css
@@ -1,16 +1,0 @@
-:host {
-    --lumo-base-color: hsl(214, 35%, 21%);
-    --lumo-primary-text-color: hsl(214, 100%, 80%);
-    --lumo-primary-color: hsl(214, 90%, 52%);
-    --lumo-primary-contrast-color: hsl(214, 100%, 98%);
-    --lumo-body-text-color: hsl(214, 25%, 90%);
-    --lumo-border-color: hsla(214, 35%, 15%, 0.5);
-    --lumo-background-color: hsl(214, 35%, 15%);
-    --lumo-shade-10pct: hsla(214, 35%, 15%, 0.1);
-    --lumo-shade-20pct: hsla(214, 35%, 15%, 0.2);
-}
-
-.dark-theme {
-    background-color: var(--lumo-background-color);
-    color: var(--lumo-primary-text-color);
-}
